% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_db.R
\name{connect_db}
\alias{connect_db}
\alias{connect_db2}
\title{Connect to PostgreSQL database}
\usage{
connect_db(
  dbname = "",
  host = "localhost",
  port = "5432",
  user = "",
  password = "",
  ...
)

connect_db2(
  dbname = "",
  host = "localhost",
  port = "5432",
  user = "",
  password = "",
  ...
)
}
\arguments{
\item{dbname, host, port, user, password}{Each a character value passed to
\code{\link[DBI:dbConnect]{DBI::dbConnect()}}. They can also be modified in the prompt.}

\item{...}{Further arguments passed to \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}
}
\description{
This function opens a prompt to insert connection details to a PostgreSQL
database. Alternatively the package \code{RPostgres} (\code{connect_db()}) or
\code{RPostgreSQL} (\code{connect_db2()}) can be used.
}
\references{
\preformatted{http://r.789695.n4.nabble.com/tkentry-that-exits-after-RETURN-tt854721.html#none
https://gist.github.com/mages/2aed2a053e355e3bfe7c#file-getlogindetails-r
Dalgaard (2001)
}
}
\author{
Miguel Alvarez
}
