% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisrepos-class.R
\docType{class}
\name{gisrepos-class}
\alias{gisrepos-class}
\title{An S4 class containing GIS-repositories}
\description{
This class contains the list of repositories, respective links and categories
at slot \strong{tables} plus their relationships at slot \strong{keys} (analogous
to foreign keys).
}
\section{Slots}{

\describe{
\item{\code{tables}}{A list containing three data frames, namely \strong{data_sets}
(main table), \strong{links} and \strong{keywords} (categories).}

\item{\code{keys}}{A list containing two data frames relating these three tables,
namely \strong{fk_links} and \strong{fk_keywords}.}
}}

\author{
Miguel Alvarez
}
